slava_talk/requirements.txt
streamlit
PyMuPDF
openai>=1.12.0
requests
beautifulsoup4
PyYAML
